Why movement is choppy

+++ src/renderer/Renderer.cpp
@@ -61,8 +61,8 @@ void Renderer::draw_from_sheet(GPU_Image *sheet, vec2 pos, Rect l_point) {
   src.h = l_point.h;
 
   GPU_Rect dst;
-  dst.x = (pos.x);
-  dst.y = (float)pos.y;
+  dst.x = static_cast<int>(pos.x);
+  dst.y = static_cast<int>(pos.y);
   dst.w = src.w;
   dst.h = src.h;
